requirements.txt
requests
python-decouple
beautifulsoup4